* master: (221 commits)
  Comment
  Comments
  msys2 build scripts: just use -Ofast
  Always prefer vararrays to alloca()
  Don't redefine alloca
  Return -1 if the scalar is 0 in crypto_scalarmult_ed25519()
  Remove trailing space
  Import the regen-msvc script
  Update ChangeLog
  Do not wipe the workspace after argon2 completes
  Remove tests for deprecated functions
  Give the compiler a change to inline index_alpha()
  Immediately allocate all required memory in argon2/scrypt
  Check reduced-round salsa variants in non-minimal mode
  Coverage exclusion -- this is just an extra, redundant check
  Remove incorrect and useless cast
  -fomit-frame-pointer makes a different on mips
  WebAssembly benefits from the 128-bit types, too
  Remove extra spaces
  Undef devel
  ...

+++ test/default/secretbox_easy2.c
@@ -22,7 +22,7 @@ main(void)
     k     = (unsigned char *) sodium_malloc(crypto_secretbox_KEYBYTES);
     mac   = (unsigned char *) sodium_malloc(crypto_secretbox_MACBYTES);
     crypto_secretbox_keygen(k);
-    randombytes_buf(m, (unsigned long long) mlen);
+    randombytes_buf(m, mlen);
     randombytes_buf(nonce, crypto_secretbox_NONCEBYTES);
     crypto_secretbox_easy(c, m, (unsigned long long) mlen, nonce, k);
     if (crypto_secretbox_open_easy(
